refactor(): change gem name to ActiveadminAddons
