+++ .gitignore
@@ -7,3 +7,6 @@ spec/dummy/log/*.log
 spec/dummy/tmp/
 spec/dummy/.sass-cache
 .gem
+.gemtags
+.tags
+.tags_sorted_by_file
