+++ Rakefile
@@ -8,14 +8,11 @@ require 'rdoc/task'
 
 RDoc::Task.new(:rdoc) do |rdoc|
   rdoc.rdoc_dir = 'rdoc'
-  rdoc.title    = 'ActiveadminUiAddons'
+  rdoc.title    = 'ActiveadminAddons'
   rdoc.options << '--line-numbers'
   rdoc.rdoc_files.include('README.rdoc')
   rdoc.rdoc_files.include('lib/**/*.rb')
 end
-
-
-
 
 Bundler::GemHelper.install_tasks
