+++ app/assets/javascripts/activeadmin_addons/all.js.coffee
@@ -1,0 +1,1 @@
+#= require ./select2
