--- app/assets/javascripts/activeadmin_ui_addons/all.js.coffee
@@ -1,1 +1,0 @@
-#= require ./select2
