+++ app/assets/stylesheets/activeadmin_addons/all.css.scss
@@ -1,0 +1,1 @@
+@import "select2";
