--- app/assets/stylesheets/activeadmin_ui_addons/all.css.scss
@@ -1,1 +1,0 @@
-@import "select2";
