+++ lib/activeadmin_addons.rb
@@ -1,0 +1,4 @@
+module ActiveadminAddons
+end
+
+require_relative './activeadmin_addons/engine'
