+++ lib/activeadmin_addons/engine.rb
@@ -1,0 +1,11 @@
+module ActiveAdminAddons
+  module Rails
+    class Engine < ::Rails::Engine
+      require 'select2-rails'
+      initializer "set boolean values addon" do |app|
+        require_relative './addons/bool_values'
+        app.config.assets.precompile +=  %w(select.scss)
+      end
+    end
+  end
+end
