+++ lib/activeadmin_addons/version.rb
@@ -1,0 +1,3 @@
+module ActiveadminAddons
+  VERSION = "0.0.2"
+end
