--- lib/activeadmin_ui_addons.rb
@@ -1,4 +1,0 @@
-module ActiveadminUiAddons
-end
-
-require_relative './activeadmin_ui_addons/engine'
