--- lib/activeadmin_ui_addons/version.rb
@@ -1,3 +1,0 @@
-module ActiveadminUiAddons
-  VERSION = "0.0.2"
-end
