+++ lib/tasks/activeadmin_addons_tasks.rake
@@ -1,0 +1,4 @@
+# desc "Explaining what the task does"
+# task :activeadmin_addons do
+#   # Task goes here
+# end
