--- lib/tasks/activeadmin_ui_addons_tasks.rake
@@ -1,4 +1,0 @@
-# desc "Explaining what the task does"
-# task :activeadmin_ui_addons do
-#   # Task goes here
-# end
