+++ spec/dummy/app/assets/javascripts/active_admin.js.coffee
@@ -1,2 +1,2 @@
 #= require active_admin/base
-#= require activeadmin_ui_addons/all
+#= require activeadmin_addons/all
