+++ spec/dummy/app/assets/stylesheets/active_admin.css.scss
@@ -1,4 +1,4 @@
-//= require activeadmin_ui_addons/all
+//= require activeadmin_addons/all
 
 // SASS variable overrides must be declared before loading up Active Admin's styles.
 //
